requirements.txt
# requirements.txt
# This file is intentionally left blank.

Pillow>=10.0.0